+++ Koha/Plugin/Com/PTFSEurope/Crontab.pm
@@ -43,7 +43,7 @@ sub tool {
 
     my $template = $self->get_template( { file => 'crontab.tt' } );
 
-    my $ct = new Config::Crontab;
+    my $ct = Config::Crontab->new();
     $ct->mode('block');
     $ct->read or do {
         $template->param( error => $ct->error );
@@ -51,19 +51,115 @@ sub tool {
         return;
     };
 
-    my $block_no = 0;
     my $blocks = [];
+    my @environment;
     for my $block ( $ct->blocks ) {
-        $block_no++;
 
+        # Get block parts
         my @comments = $block->select( -type => 'comment' );
         my @env      = $block->select( -type => 'env' );
         my @events   = $block->select( -type => 'event' );
 
-        push @{$blocks}, { no => $block_no,  comments => \@comments, environment => \@env, events => \@events };
+        # Get block id
+        my $id;
+        my $id_line = shift @comments;
+        if ( $id_line->data =~ /# BLOCKID: (\d+)/ ) {
+            $id = $1;
+        } else {
+            $template->param( error => "Found block with missing ID: " . $id_line->data );
+            $self->output_html( $template->output() );
+            return;
+        }
+
+        # Skip first block (plugin header)
+        next if ($id == 0);
+
+        # Global environment block
+        unless (@events) {
+            push @environment, @env;
+            next;
+        }
+
+        push @{$blocks},
+          {
+            id          => $id,
+            comments    => \@comments,
+            environment => \@env,
+            events      => \@events
+          };
     }
-    $template->param( blocks => $blocks );
+    $template->param(
+        environment => \@environment,
+        blocks      => $blocks
+    );
     $self->output_html( $template->output() );
+}
+
+sub api_routes {
+    my ( $self, $args ) = @_;
+
+    my $spec_str = $self->mbf_read('openapi.json');
+    my $spec     = decode_json($spec_str);
+
+    return $spec;
+}
+
+sub api_namespace {
+    my ($self) = @_;
+
+    return 'crontab';
+}
+
+sub install() {
+    my ( $self, $args ) = @_;
+
+    my $ct = Config::Crontab->new();
+    $ct->mode('block');
+    $ct->read or do {
+        return 0;
+    };
+
+    # We should take a backup here before we do anything else
+
+    # Read existing crontab, update it to identify blocks
+    # so we can manage that we can recognise for management
+    # BLOCKID:
+    my $block_id = 1;
+    for my $block ( $ct->blocks ) {
+        for my $comment (
+            $block->select(
+                -type    => 'comment',
+                -data_re => 'Koha Crontab manager'
+            )
+          )
+        {
+            return 0;    # Already installed
+        }
+        $block->first(
+            Config::Crontab::Comment->new(
+                -data => "# BLOCKID: " . $block_id++
+            )
+        );
+    }
+
+    # Set first block to state we're maintained by the plugin
+    my $block = Config::Crontab::Block->new();
+    $block->first(
+        Config::Crontab::Comment->new(
+            -data =>
+"# This crontab file is managed by the Koha Crontab manager plugin"
+        )
+    );
+    $block->last( Config::Crontab::Comment->new( -data => "# BLOCKID: 0" ) );
+    $ct->first($block);
+
+    # Add a hash so we can tell if the managed content has
+    # been modified externally and warn the user during updates
+
+    ## write out crontab file
+    $ct->write;
+
+    return 1;
 }
 
 1;