+++ CHANGELOG.md
@@ -10,5 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add beggining of management page as tool
 - Testing Config::Crontab for loading instance crontab
 - Build start of UI using [crontab-ui](https://github.com/alseambusher/crontab-ui/tree/main) for inspiration
+- Add install method to parse and manipulate crontab for the first time
+- Add API endpoint for updateing cron lines